+++ src/routes/scholar/[id]/+page.ts
@@ -9,13 +9,13 @@ export const load: PageLoad = async ({ parent, params }) => {
 	const { data: scholar } = await supabase.from('scholars').select().eq('id', scholarID).single();
 
 	// Get the scholar's commitments
-	const { data: commitments } = await supabase
+	const { data: volunteers } = await supabase
 		.from('volunteers')
-		.select('*, roles(name, venueid)')
+		.select('*, roles(name, venueid, approver)')
 		.eq('scholarid', scholarID);
 
-	const venueids = commitments
-		? commitments.map((c) => c.roles?.venueid).filter((v) => v !== undefined)
+	const venueids = volunteers
+		? volunteers.map((c) => c.roles?.venueid).filter((v) => v !== undefined)
 		: [];
 
 	const { data: venues } =
@@ -70,7 +70,7 @@ export const load: PageLoad = async ({ parent, params }) => {
 		.contains('authors', [scholarID]);
 	if (submissionsError) console.log(submissionsError);
 
-	// Get the scholar's reviewing assignments
+	// Get the scholar's approved reviewing assignments
 	const { data: reviews, error: reviewsError } = await supabase
 		.from('assignments')
 		.select('*, submissions(*)')
@@ -79,9 +79,24 @@ export const load: PageLoad = async ({ parent, params }) => {
 		.eq('approved', true);
 	if (reviewsError) console.log(reviewsError);
 
+	// Get the roles for which the scholar is the role approver.
+	const { data: approver, error: approverError } = await supabase
+		.from('roles')
+		.select('*')
+		.in('approver', reviews?.map((c) => c.role) || []);
+	if (approverError) console.log(approverError);
+
+	// Get the assignments for which the scholar is the role approver, to show in the scholar's dashboard.
+	const { data: approvals, error: approvalsError } = await supabase
+		.from('assignments')
+		.select('*, scholars(*), submissions(*)')
+		.in('role', approver?.map((r) => r.id) || [])
+		.eq('approved', false);
+	if (approvalsError) console.log(approvalsError);
+
 	return {
 		scholar,
-		commitments,
+		commitments: volunteers,
 		venues,
 		editing,
 		tokens: tokens,
@@ -90,6 +105,7 @@ export const load: PageLoad = async ({ parent, params }) => {
 		currencies: currencies,
 		minting: minting,
 		pending: pending,
-		reviews: reviews
+		reviews: reviews,
+		approvals: approvals
 	};
 };